% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_terminus.R
\name{add_node_terminus}
\alias{add_node_terminus}
\title{Add attribute node terminus}
\usage{
add_node_terminus(graph, check_node_attributes = FALSE)
}
\arguments{
\item{graph}{A tidygraph or igraph object}

\item{check_node_attributes}{Logical value that determines whether the node
attributes are checked for an existing \code{node_terminus} attribute.}
}
\value{
The function returns the graph with a new node attribute called
\code{node_terminus}. It has the values \code{start} and \code{end} for the terminal nodes
in a graph, \code{fusion} and \code{fission} for merges and splits, and \code{internal} for
linearily connected nodes.
}
\description{
Adds a node attribute called \code{node_terminus} to the nodes of a graph.
For tracing organelles, fusion and fission events are interesting to record
as well as the start and end position of the organelles. These can be
identified based on the degrees of the nodes.
}
