% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_angles.R
\name{angular_mean}
\alias{angular_mean}
\title{Mean angle}
\usage{
angular_mean(x, units = "radians")
}
\arguments{
\item{x}{Numeric vector}

\item{units}{A character string declaring the units of the angles. The
default is \code{'radians'} but can also be \code{'degree'}.}
}
\value{
The average angle in radians or degrees.
}
\description{
Circular angles repeat every 2*pi radians which has to be taken into account
for calculating the averages over angles.
}
