% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_angular_range.R
\name{angular_range_test}
\alias{angular_range_test}
\title{Angular range test}
\usage{
angular_range_test(x, angular_range = c(0, 2 * pi))
}
\arguments{
\item{x}{Numeric vector with angles to be tested.}

\item{angular_range}{Numeric vector with only the limits of
the angular test range. The full circle range in radians, \code{anglular_range=c(0, 2*pi)},
is chosen by default.}
}
\value{
A logical vector of same length as x with values \code{TRUE} if the
individual angle is within the test range or \code{FALSE} if not.
}
\description{
This function tests if angles are within a specified range. The angles \code{x}
and the test range \code{angular_range} need to be in the same units but can be
any circular revolution of the angles ( \code{a = a + 2*pi} ).
}
