% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tracks.R
\name{bind_tracks}
\alias{bind_tracks}
\title{Bind a list of track graphs together}
\usage{
bind_tracks(graph_list)
}
\arguments{
\item{graph_list}{A list with an graph object in each element. The graphs
need to be of class \code{'igraph'} or \code{'tidygraph'}.}
}
\value{
A single combined graph of class \code{'igraph'} and\code{'tbl_graph'}
compatible with the tidygraph package.
}
\description{
This is the reverse of the \code{split_by_tracks} function. It binds a list of
graphs together, e.g. to join multiple tracks into one graph. The individual
tracks do not need to be connected (share nodes).
}
