% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_identity_layout.R
\name{create_identity_layout}
\alias{create_identity_layout}
\title{Create identity layout}
\usage{
create_identity_layout(graph)
}
\arguments{
\item{graph}{An graph object from igraph or tidygraph packages.
The coordinates must be stored in node attributes called \code{node_x} and
\code{node_y}.
The edges must have identifiers for source and target in attributes
\code{source_id} and \code{target_id} to repair \code{from} and \code{to} in subgraphs.}
}
\value{
A graph layout of class 'layout_ggraph' and 'layout_tbl_graph'
suitable for plotting with the ggraph package.
}
\description{
Create ggraph layout from graph to plot the original x,y coordinates.
What you would expect from \code{ggraph::create_layout(graph, layout = "identity")}.
It repairs the edge attributes \code{from} and \code{to} in subgraphs.
}
