% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_branches.R
\name{find_branches}
\alias{find_branches}
\title{Detecting branches}
\usage{
find_branches(graph)
}
\arguments{
\item{graph}{A tidygraph or igraph object.}
}
\value{
The function returns the graph in which the branches have been labeled. The
identifier for the branches is stored in a new edge attribute \code{branch_id}.
}
\description{
This function detects all the branches of a graph. It takes a graph and returns a list of vectors, one for
each branch, containing the node IDs within the branch (including the start
and end).
}
