% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_tbl_graph_identity.R
\name{layout_tbl_graph_identity}
\alias{layout_tbl_graph_identity}
\title{Calculate identity layout}
\usage{
layout_tbl_graph_identity(graph, x, y, circular = FALSE)
}
\arguments{
\item{graph}{An graph object from igraph or tidygraph packages.}

\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{circular}{Logical vector.}
}
\value{
A graph layout of class 'layout_ggraph' and 'layout_tbl_graph'
suitable for plotting with the ggraph package.
}
\description{
Create ggraph layout from graph to plot the original x,y coordinates.
What you would expect from \code{ggraph::create_layout(graph, layout = "identity", x, y)}.
It repairs the edge attributes \code{from} and \code{to} in subgraphs.
}
