% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_angles.R
\name{nice_angles}
\alias{nice_angles}
\title{Nice angles}
\usage{
nice_angles(x, units = "radians")
}
\arguments{
\item{x}{Numeric vector}

\item{units}{A character string declaring the units of the angles. The
default is \code{'radians'} but can also be \code{'degree'}.}
}
\value{
A numeric vector with angles converted to the range \verb{[0, 2*pi[}. If
the units are in \code{'degree'}, the output is converted to the range \verb{[0, 360[}
instead.
}
\description{
Circular angles repeat every 2*pi radians and can make angular calculations
challenging. The function \code{nice_angles} converts angles to
\verb{[0, 2*pi[} or \verb{[0°, 360°[} to simplify calculations.
}
