% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_TrackMate.R
\name{read_TrackMate}
\alias{read_TrackMate}
\title{Read TrackMate save files (xml)}
\usage{
read_TrackMate(
  filepath,
  import = c("tracks", "nodes", "edges", "settings", "graph", "full")
)
}
\arguments{
\item{filepath}{Path to the xml file.}

\item{import}{Import options to select the retrieved data.
\code{'tracks'} will return a tibble with the tracks statistics.
\code{'nodes'} will return the nodeslist in a tibble.
\code{'edges'} will return the edgeslist in a tibble.
\code{'settings'} will return a list of settings from TrackMate.
\code{'graph'} will return an graph object.
\code{'full'} will return a list containing the settings, tracks, nodelist, and
edgelist.}
}
\value{
The value is a tibble with tracks, nodes (node list), or edges
(edge list) if the respective import option was selected.
The import option \code{'settings'} will return a list with the TrackMate
settings.
For full import, the function returns a list with tibbles for the tracks,
nodes, and edges, and with a list with the TrackMate settings.
If the graph was selected, a graph of classes \code{'tbl_graph'} and \code{'igraph'}
will be returned. It is compatible with the \code{igraph} and \code{tidygraph}
packages and can be plotted with the \code{ggraph} package.
}
\description{
Read xml files saved with TrackMate.
}
