% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_tracks.R
\name{split_by_tracks}
\alias{split_by_tracks}
\title{Split a graph by tracks}
\usage{
split_by_tracks(graph, tracks, nodes, edges)
}
\arguments{
\item{graph}{A graph object of class \code{'igraph'} or \code{'tbl_graph'} from the
igraph or tidygraph packages.}

\item{tracks}{A data.frame or vector with a column \code{'track_id'} that
identifies the tracks.}

\item{nodes}{A data.frame or tibble with the node list. It must start with a
column \code{'node_id'}}

\item{edges}{A data.frame or tibble with the edge list. It must contain
columns \code{'source_id'}, \code{'target_id'} to connect nodes, and a column
\code{'track_id'} to reference the tracks.}
}
\value{
A list of individual graphs for each track.
}
\description{
This function splits a graph with unconnected tracks obtained from TrackMate
into a list of graphs with the individual tracks. This is useful for some
calculations and to further investigate interesting tracks.
The function takes a single graph or a node and an edge list as input.
The tracks can also be specified optionally.
}
