% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_angles.R
\name{weighted_angular_mean}
\alias{weighted_angular_mean}
\title{Speed-weighted angular mean (Yamartino 1984)}
\usage{
weighted_angular_mean(angle, speed, units = "radians")
}
\arguments{
\item{angle}{Numeric vector with angles}

\item{units}{A character string declaring the units of the angles. The
default is \code{'radians'} but can also be \code{'degree'}.}

\item{velocity}{Numeric vector with velocities used to weigh the mean}
}
\value{
The average angle in radians or degrees.
}
\description{
Circular angles repeat every 2*pi radians which has to be taken into account
for calculating the angular means.
}
